Moved the "Entering location " for stages back 5 points (pixels? Squares, can't remember)
Moved the "End" Location for special stages to 0x3d, this is due to the 30k boss being ONE pixel behind, and thus triggering all the achievements it shouldn't.

Changed the description of Junior Kill achievements, to not mention hard difficulty

Redid a bit of the 30k boss, because it unlocks incorrectly when you fight the >30k <900k boss.   This causes changes to all Special_Boss_Values